requirements.txt
python-arango==7.9.1
Faker==24.1.0
docker==7.0.0
